[Feat] Dividie delegate and view.
